setup shared toReqRes
